Convert AttributeId to a string type.
Converting the id type to a string to avoid having to convert them to a
string each time it needs to be made human readable, such as when
converting to JSON or eventually when displaying to the user. Now, it
just has to be converted to the string constant when the message from
MakeMKV is parsed.

Also, converted the `int32` fields to plain `int` types. Using an
explicit int type was a force of habit from my day job.