Add store package.
Added the store package for storing global data and ensuring that it can
safely be read and written accross goroutines. Updated the startup to
initialize the state and the service to use it when getting the server
status. As part of this, had to update the DB logic to return the
drive's database id.